requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
mysql-connector-python==8.0.31
python-dotenv==1.0.0
